crctab: use stdint.h types

+++ crctab.c
@@ -1,6 +1,6 @@
-
+#include <stdint.h>
 /* crctab calculated by Mark G. Mendel, Network Systems Corporation */
-unsigned short crc16tab[256] = {
+uint16_t crc16tab[256] = {
     0x0000, 0x1021, 0x2042, 0x3063, 0x4084, 0x50a5, 0x60c6, 0x70e7, 0x8108,
     0x9129, 0xa14a, 0xb16b, 0xc18c, 0xd1ad, 0xe1ce, 0xf1ef, 0x1231, 0x0210,
     0x3273, 0x2252, 0x52b5, 0x4294, 0x72f7, 0x62d6, 0x9339, 0x8318, 0xb37b,
@@ -68,7 +68,7 @@ unsigned short crc16tab[256] = {
 /*     hardware you could probably optimize the shift in assembler by  */
 /*     using byte-swap instructions.                                   */
 
-long crc32tab[] =
+uint32_t crc32tab[] =
     {/* CRC polynomial 0xedb88320 */
      0x00000000, 0x77073096, 0xee0e612c, 0x990951ba, 0x076dc419, 0x706af48f,
      0xe963a535, 0x9e6495a3, 0x0edb8832, 0x79dcb8a4, 0xe0d5e91e, 0x97d2d988,
